requirements.txt
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
Django==3.1.7
django-extensions==3.1.1
djangorestframework==3.12.2
idna==2.10
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.17
ptyprocess==0.7.0
Pygments==2.8.1
pymarc==4.0.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
